타입밖으로뻄
