+++ src/types/toolTipType.ts
@@ -1,0 +1,10 @@
+export type LocationType = {
+  x: number
+  y: number
+  width: number
+  height: number
+  top: number
+  right: number
+  bottom: number
+  left: number
+}
